chore: documentation example code was not secure
The example code shown was not secure. While this is not a vulnerability since it is not used as actual code, developers should not be given unsafe code as documentation and expect it to be the norm. This PR fixes it to be secure of TOCTOU race conditions, as before it would have been possible to, after io.mkdir(), another process quickly changes the directory to be a symlink, granting the arbitrary process to abuse the following fs_write() call to make the Nyxian example code write to any file it wants to.